insight_token/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
spacy==3.8.4
google-cloud-storage==2.19.0
